+++ ptl_bolts/models/VAEs/__init__.py
@@ -1,0 +1,1 @@
+from ptl_bolts.models.VAEs.basic.vae import VAE
