--- ptl_bolts/models/variational_autoencoders/__init__.py
@@ -1,1 +1,0 @@
-from ptl_bolts.models.variational_autoencoders.basic.vae import VAE
